+++ src/decomposition.rs
@@ -1,7 +1,4 @@
 //! Matrix decompositions.
 
-mod lu;
-mod qr;
-
-pub use lu::LUFactorized;
-pub use qr::QRFactorized;
+pub mod lu;
+pub mod qr;