+++ src/equation.rs
@@ -1,6 +1,6 @@
 //! Matrix equation solvers.
 
-use crate::decomposition::LUFactorized;
+use crate::decomposition::lu;
 use crate::{InvalidInput, Real, Scalar};
 use ndarray::{Array1, ArrayBase, Data, Ix1, Ix2};
 
@@ -50,7 +50,7 @@ where
             a.nrows()
         )));
     }
-    let factorized = LUFactorized::from(a.to_owned());
+    let factorized = lu::Factorized::from(a.to_owned());
     if factorized.is_singular() {
         Err(InvalidInput::Value("`a` is a singular matrix".to_string()))
     } else {